backend/requirements.txt
fastapi
uvicorn
pandas
spacy
python-multipart
PyPDF2 
python-docx